Split VPS config modules
